protoc-gen-go: simplify init function generation
We generate a single init function containing registration calls for
all the things in a .pb.go that need one (messages, extensions, etc.)

Some of these registrations are added to a list as the generator processes
the type that needs to be registered (e.g., messages.) Some are added at
the end while generating the init function (e.g., enums). Extensions
are done in both places.

Make this consistent and always generate the registration functions
during processing. This simplifies the code a bit, and makes it easier
for a new generator to be consistent with its output.

No changes to the number of init functions, just the order of
registration calls in the existing one.