go.mod: rely on go get support for modules instead of local paths
In CL 136736, we renamed the import path as github.com/golang/protobuf/v2
so that it can be fetched as a module with "go get". Use that mechanism to
introduce the dependency on v2 instead of using the replace feature to point
at a local checkout.

While we are at it, also fix Travis:
* The test.bash script is now an exact copy of the one in v2,
except that we do not run "./internal/cmd/generate-types".
* The .travis.yml config is now an exact copy of the one in the v2
where it simply delegates dependency management and testing to test.bash.

The v2 dependency can be updated to use the latest by running:
	go get github.com/golang/protobuf/v2@api-v2

Change-Id: I2284d0bae82afc5306c91977aa44d09f7e5a41cc
Reviewed-on: https://go-review.googlesource.com/136737
Reviewed-by: Damien Neil <dneil@google.com>