cmd/protoc-gen-go: factor out oneof field naming
A oneof is represented by a single struct field of interface type.
Pull out the decision of what to name that field into a separate
function.

The function is trivial (return oneof.GoName), but factoring out the
field name like this makes it a bit easier to experiment with changes to
the oneof implementation in the future.

Change-Id: I1114b68c85cb6608852fa1c6bf4103ff58fd5de6
Reviewed-on: https://go-review.googlesource.com/c/146397
Reviewed-by: Joe Tsai <thebrokentoaster@gmail.com>