protoapi: move registration from proto to protoapi
In order to generate the descriptor package with reflection support,
it must stop depending on the v1 proto package itself.
However, it still depends on v1 proto for registration.

Thus, to break descriptor's dependency on v1 proto, we temporarily
move registration to the protoapi package.
This move is temporary since registration will eventually be held in the
v2 package and newly generated messages will directly register with the
v2 registries, and legacy messages will continue to register through the
v1 proto package, which will just forward the registrations to v2.

However, the v2 registry is not complete yet.

Change-Id: Iba16d5b3a947c332fc0506217e9e033076764092
Reviewed-on: https://go-review.googlesource.com/c/152546
Reviewed-by: Damien Neil <dneil@google.com>